Feature/test

+++ contracts/SafeMathLib.sol
@@ -12,20 +12,20 @@ pragma solidity ^0.4.23;
  */
 library SafeMathLib {
 
-    function times(uint a, uint b)
+    function times(uint256 a, uint256 b)
         internal pure returns (uint) {
         uint c = a * b;
         assert(a == 0 || c / a == b);
         return c;
     }
 
-    function minus(uint a, uint b)
+    function minus(uint256 a, uint256 b)
         internal pure returns (uint) {
         assert(b <= a);
         return a - b;
     }
 
-    function plus(uint a, uint b)
+    function plus(uint256 a, uint256 b)
         internal pure returns (uint) {
         uint c = a + b;
         assert(c>=a);
